# cmd/chat-bots-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bot Factory backend API written in Go Gin",
        "title": "Bot Factory API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/chat-bot/{chat-bot-id}/file/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Saves a file to database and adds it to the chat-bot",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Saves a file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat Bot ID",
                        "name": "chat-bot-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SaveFileResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chat-bot/{chat-bot-id}/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return chat-bot files meta-data from database",
                "tags": [
                    "file"
                ],
                "summary": "Get chat-bot files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat Bot ID",
                        "name": "chat-bot-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ChatBotFiles"
                        }
                    }
                }
            }
        },
...
        "/api/v1/file/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a file meta-data by its ID from the database",
                "tags": [
                    "file"
                ],
                "summary": "Returns a file meta-data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.File"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a file from database and chat-bot",
                "tags": [
                    "file"
                ],
                "summary": "Removes a file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.RemoveFileResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/login/": {
            "post": {
                "description": "Login user with email and password",
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/register/": {
            "post": {
                "description": "Register user with email and password",
                "tags": [
                    "auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Register credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch user data from database by session token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Fetches user by token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user plan (Free: \"free_plan\", Business: \"business_plan\", Enterprise: \"enterprise_plan\")",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user plan",
                "parameters": [
                    {
                        "description": "Chat-bot ID",
                        "name": "new-plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdatePlanResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ChatBotFiles": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.File"
                    }
                }
            }
        },
        "schemas.File": {
            "type": "object",
            "properties": {
                "chat_bot_id": {
                    "type": "integer"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "schemas.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "schemas.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "schemas.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "schemas.RegisterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.RemoveFileResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "schemas.SaveFileResponse": {
            "type": "object",
            "properties": {
                "file_id": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.UpdatePlanRequest": {
            "type": "object",
            "properties": {
                "plan": {
                    "type": "string"
                }
            }
        },
        "schemas.UpdatePlanResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "schemas.User": {
            "type": "object",
            "properties": {
                "bots_left": {
                    "type": "integer"
                },
                "bytes_data_left": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "messages_left": {
                    "type": "integer"
                },
                "plan": {
                    "type": "string"
                },
                "plan_bought_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "cookie"
        }
    }
}